third_party/blink/web_tests/platform/linux/external/wpt/mathml/presentation-markup/operators/stretchy-largeop-with-default-font-2-expected.txt

This is a testharness.js-based test.
[FAIL] Test that non-stretchy largeops are bigger in display mode.
  assert_greater_than: width of largeop '⨀' expected a number greater than 16 but got 16
Harness: the test ran to completion.
